update CRUD user

+++ nw1-api-sas/src/main/java/sit/int222/nw1apisas/services/UserService.java
@@ -1,8 +1,9 @@
 package sit.int222.nw1apisas.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.server.ResponseStatusException;
 import sit.int222.nw1apisas.dtos.CreateUpdateUserDto;
 import sit.int222.nw1apisas.entities.User;
 import sit.int222.nw1apisas.exceptions.ItemNotFoundException;
@@ -19,7 +20,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User createUser(@RequestBody CreateUpdateUserDto newUser) {
+    public User createUser(CreateUpdateUserDto newUser) {
         User user = new User();
         user.setUsername(newUser.getUsername());
         user.setName(newUser.getName());
@@ -44,9 +45,13 @@ public class UserService {
         return userRepository.saveAndFlush(existingUser);
     }
 
-    public void deleteUser(Integer id){
-        userRepository.findById(id).orElseThrow(() -> new ItemNotFoundException("The announcement is not found."));
-        userRepository.deleteById(id);
+    public void deleteUser(Integer id) {
+        if (id == null && !(id instanceof Integer)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "id must be integer");
+        } else {
+            userRepository.findById(id).orElseThrow(() -> new ItemNotFoundException("The announcement is not found."));
+            userRepository.deleteById(id);
+        }
     }
 
 }
